+++ front/app/containers/Admin/pagesAndMenu/containers/ContentBuilder/components/Widgets/_shared/AdminPubsCarrousel/AdminPubCard/index.tsx
@@ -75,7 +75,7 @@ export const AdminPubCard = ({
             <Icon
               name="folder-solid"
               height="20px"
-              ml="-4px"
+              ml="-2px"
               mr="4px"
               mt="0px"
             />
